위 방향 요리 재료 표기 버그 수정,  현재 방 재료 표기

+++ Assets/Scripts/System/MapManager.cs
@@ -147,7 +147,7 @@ public class MapManager : Singleton<MapManager>
             leftChunkText.text = chunkItem.data.StringChunkType;
             leftChunkDropTableUI.SetDropItemImages(chunkItem.data.dropTable);
         }
-        else
+        else if ((moveChunkPos - playerChunkPos) == Vector2Int.up)
         {
             upChunkText.text = chunkItem.data.StringChunkType;
             upChunkDropTableUI.SetDropItemImages(chunkItem.data.dropTable);
